+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/data/base/BaseRecipeProvider.java
@@ -5,9 +5,12 @@ import com.christofmeg.brutalharvest.common.init.ItemRegistry;
 import com.christofmeg.brutalharvest.common.init.TagRegistry;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.*;
+import net.minecraft.data.recipes.packs.VanillaRecipeProvider;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.AbstractCookingRecipe;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.item.crafting.RecipeSerializer;
@@ -15,12 +18,14 @@ import net.minecraft.world.level.ItemLike;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.common.crafting.conditions.IConditionBuilder;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
+import java.lang.reflect.Field;
+import java.util.*;
 import java.util.function.Consumer;
 
 @SuppressWarnings("unused")
-public class BaseRecipeProvider extends RecipeProvider implements IConditionBuilder {
+public class BaseRecipeProvider extends VanillaRecipeProvider implements IConditionBuilder {
 
     public BaseRecipeProvider(PackOutput packOutput) {
         super(packOutput);
@@ -62,6 +67,29 @@ public class BaseRecipeProvider extends RecipeProvider implements IConditionBuil
                 .requires(requires)
                 .unlockedBy("fabric", has(ItemRegistry.FABRIC.get()))
                 .save(consumer, modLoc(getItemName(item)));
+    }
+
+    protected void woolRecipeBuilder(TagKey<Item> ingredient, Item banner, Item bed, Item carpet, Consumer<FinishedRecipe> consumer) {
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, banner)
+                .define('#', ingredient)
+                .define('|', Tags.Items.RODS_WOODEN)
+                .pattern("###")
+                .pattern("###")
+                .pattern(" | ")
+                .unlockedBy("has_wool", has(ItemTags.WOOL))
+                .save(consumer, modLoc(getItemName(banner)));
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, bed)
+                .define('#', ingredient)
+                .define('X', ItemTags.PLANKS)
+                .pattern("###")
+                .pattern("XXX")
+                .unlockedBy("has_wool", has(ItemTags.WOOL))
+                .save(consumer, modLoc(getItemName(bed)));
+        ShapedRecipeBuilder.shaped(RecipeCategory.MISC, carpet)
+                .define('#', ingredient)
+                .pattern("##")
+                .unlockedBy("has_wool", has(ItemTags.WOOL))
+                .save(consumer, modLoc(getItemName(carpet)));
     }
 
     protected static void blasting(List<ItemLike> ingredients, RecipeCategory category, ItemLike result, float experience, int cookingTime, String group, Consumer<FinishedRecipe> consumer) {